Combine with declarative recipes for complete migration
